refactor: Refactor models to inherit from new base Model class
This commit change enhances the codebase by introducing
a unified base class for model entities, simplifying
import statements and making the overall structure more
maintainable and consistent.
