+++ .flake8
@@ -3,6 +3,7 @@ exclude =
     .git
     __pycache__
     .venv
+    pytrello2/models/__init__.py
 ignore = E129, E402, F841, W504, E128
 max-line-length = 88
 extend-ignore = E203
